Refactored LevelBuilder, Level and LevelFactory to reduce severity of "Data class" warning by inCode <organisation>
