+++ src/main/java/doob/model/Level.java
@@ -2,6 +2,7 @@ package doob.model;
 
 import doob.DLog;
 import doob.level.CollisionManager;
+import doob.level.CollisionResolver;
 import doob.level.LevelObserver;
 import doob.level.PowerUpManager;
 import doob.model.powerup.PowerUp;
@@ -16,6 +17,8 @@ import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 /**
@@ -442,6 +445,101 @@ public class Level {
                     players.get(0).setSpeed(0);
                     break;
             }
+        }
+    }
+
+    /**
+     * Class that assists in building level.
+     */
+    public static class Builder {
+
+        private Canvas canvas;
+        private ArrayList<Ball> balls;
+        private ArrayList<Player> players;
+        private ArrayList<Wall> walls;
+        private int time;
+
+        /**
+         * Canvas setter.
+         * @param canvas canvas
+         */
+        public Builder setCanvas(Canvas canvas) {
+            this.canvas = canvas;
+            return this;
+        }
+
+        /**
+         * Time setter.
+         * @param time time
+         */
+        public Builder setTime(int time) {
+            this.time = time;
+            return this;
+        }
+
+        /**
+         * Balls setter.
+         * @param balls balls
+         */
+        public Builder setBalls(ArrayList<Ball> balls) {
+            this.balls = balls;
+            return this;
+        }
+
+        /**
+         * Walls setter.
+         * @param walls walls
+         */
+        public Builder setWalls(ArrayList<Wall> walls) {
+            this.walls = walls;
+            return this;
+        }
+
+        /**
+         * Player setter.
+         * @param players players
+         */
+        public Builder setPlayers(ArrayList<Player> players) {
+            this.players = players;
+            return this;
+        }
+
+        /**
+         * Builds the level.
+         * @return level
+         */
+        public Level build() {
+            Level level = new Level(canvas);
+            Wall right = new Wall((int) canvas.getWidth(), 0, 1, (int) canvas.getHeight());
+            Wall left = new Wall(0, 0, 1, (int) canvas.getHeight());
+            Wall ceiling = new Wall(0, 0, (int) canvas.getWidth(), 1);
+            Wall floor = new Wall(0, (int) canvas.getHeight(), (int) canvas.getWidth(), 1);
+
+            level.setLeft(left);
+            level.setRight(right);
+            level.setCeiling(ceiling);
+            level.setFloor(floor);
+
+            Collections.sort(walls, new Comparator<Wall>() {
+                @Override
+                public int compare(Wall w1, Wall w2) {
+                    return Integer.compare(w1.getX(), w2.getX());
+                }
+            });
+
+            walls.add(right);
+            walls.add(0, left);
+            walls.add(floor);
+            walls.add(ceiling);
+
+            level.setCollisionManager(new CollisionManager(level, new CollisionResolver(level)));
+            level.setPowerUpManager(new PowerUpManager(level));
+            level.setBalls(balls);
+            level.setPlayers(players);
+            level.setWalls(walls);
+            level.setTime(time);
+            level.setCurrentTime(time);
+            return level;
         }
     }
